+++ examples/fragments/README.md
@@ -1,0 +1,13 @@
+# Fragments Example
+
+## Concepts
+
+The example uses [fragments](https://yew.rs/docs/en/concepts/html/lists/#fragments) to have multiple root elements in the `html!` macro.
+
+## Improvements
+
+The example doesn't really make use of fragments in a way that requires them.
+The same functionality and visual result can be accomplished without the use of fragments.
+
+This example should use fragments for something that can't be easily replicated without them or be removed entirely.
+Other examples already make use of fragments.
